+++ src/PhpBrew/Command/InstallCommand.php
@@ -16,13 +16,14 @@ use PhpBrew\Tasks\DSymTask;
 use PhpBrew\VariantParser;
 use PhpBrew\Build;
 use PhpBrew\DirectorySwitch;
+use CLIFramework\Command;
 
 
 /*
  * TODO: refactor tasks to Task class.
  */
 
-class InstallCommand extends \CLIFramework\Command
+class InstallCommand extends Command 
 {
     public function brief() { return 'install php'; }
 
@@ -169,23 +170,30 @@ class InstallCommand extends \CLIFramework\Command
         $dsym = new DSymTask($this->logger);
         $dsym->patch( $build, $this->options );
 
-
-        $phpConfigFile = $this->options->production ? 'php.ini-production' : 'php.ini-development';
-        $this->logger->info("---> Copying $phpConfigFile ");
-        if ( file_exists($phpConfigFile) ) {
-            if ( ! file_exists( Config::getVersionEtcPath($version) ) ) {
-                mkdir( Config::getVersionEtcPath($version) , 0755 , true );
+        // copy php-fpm config
+        $this->logger->info("---> Creating php-fpm.conf");
+        $phpfpmConfigPath = "sapi/fpm/php-fpm.conf";
+        $phpfpmTargetConfigPath = Config::getVersionEtcPath($version) . DIRECTORY_SEPARATOR . 'php-fpm.conf';
+        if ( file_exists($phpfpmConfigPath) ) {
+            if ( ! file_exists( $phpfpmTargetConfigPath ) ) {
+                copy($phpfpmConfigPath, $phpfpmTargetConfigPath);
+            } else {
+                $this->logger->notice("Found existing $phpfpmTargetConfigPath.");
             }
+        }
 
+
+        $phpConfigPath = $options->production ? 'php.ini-production' : 'php.ini-development';
+        $this->logger->info("---> Copying $phpConfigPath ");
+        if( file_exists($phpConfigPath) ) {
             $targetConfigPath = Config::getVersionEtcPath($version) . DIRECTORY_SEPARATOR . 'php.ini';
-            if ( file_exists($targetConfigPath) ) {
-                $this->logger->notice("$targetConfigPath exists, do not overwrite.");
+            if( file_exists($targetConfigPath) ) {
+                $this->logger->notice("Found existing $targetConfigPath.");
             } else {
 
                 // TODO: Move this to PhpConfigPatchTask
-
                 // move config file to target location
-                rename( $phpConfigFile , $targetConfigPath );
+                copy( $phpConfigPath , $targetConfigPath );
 
                 // replace current timezone
                 $timezone = ini_get('date.timezone');