+++ src/PhpBrew/Command/OffCommand.php
@@ -1,0 +1,7 @@
+<?php
+namespace PhpBrew\Command;
+
+class OffCommand extends VirtualCommand
+{
+    public function brief() { return 'go back to the system php'; }
+}